Unset itemId when trigger type is not "product"
Because of a bug on Snipcart's side we need to remove any
itemId from the request if the trigger type is not set to product.
If any itemId is sent along the trigger type will be ignored and
the coupon will be applied as soon as this product will be added to
the cart.